pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "gnn3"
version = "0.1.0"
description = "Packet-switched GNN routing research stack with Packet-Mamba3 variants"
readme = "README.md"
requires-python = ">=3.12,<3.13"
authors = [{ name = "Codex" }]
dependencies = [
    "einops>=0.8.2",
    "matplotlib>=3.10.8",
    "networkx>=3.6.1",
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "pymupdf4llm>=1.27.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "rich>=14.3.3",
    "scipy>=1.17.1",
    "seaborn>=0.13.2",
    "tensorboard>=2.20.0",
    "tqdm>=4.67.3",
]

[dependency-groups]
dev = [
  "mypy>=1.17.1",
  "pytest>=8.4.1",
  "pytest-xdist>=3.8.0",
  "ruff>=0.13.0",
  "types-pyyaml>=6.0.12.20250915",
]

[tool.hatch.build.targets.wheel]
packages = ["src/gnn3"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-ra"

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests", "scripts"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
no_implicit_optional = true
pretty = true
packages = ["gnn3"]
